# docs/api-spec.yaml
openapi: 3.0.3
...
info:
  title: myshop API
  version: 1.0.0
  description: |
    E-commerce REST API for the myshop storefront.

    Conventions:
    - All money values are **integer cents** (`priceCents: 1900` = $19.00). Never floats.
    - All list endpoints use the shared pagination envelope
      `{ content, page, size, totalElements, totalPages }`.
    - All errors are RFC 9457 problem+json with `code` (stable machine code)
      and `traceId` (quote in bug reports).
    - Status codes: 400 validation · 401 unauthenticated · 403 forbidden ·
      404 missing · 409 conflict (e.g. out of stock) · 422 business rule · 500 unexpected.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List active products ✅
      security: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
            default: newest
      responses:
        '200':
          description: Paged product summaries, prices aggregated from variants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedProducts' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/v1/products/{slug}:
    get:
      tags: [Products]
      summary: Product detail with variants and images ✅
      security: []
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        '404':
          description: No ACTIVE product with this slug
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                code: PRODUCT_NOT_FOUND
                detail: No product with slug 'xyz'
...
components:
  parameters:
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
      description: Stable forever — changes require a redirect (AGENTS.md)
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { code: VALIDATION_ERROR }
...
  schemas:
    Problem:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          description: Stable machine-readable code
          example: PRODUCT_NOT_FOUND
        detail:
          type: string
          description: Human-readable message
        traceId:
          type: string
          description: Quote this when reporting errors
        status: { type: integer }
        instance: { type: string }
    Image:
      type: object
      required: [url, alt, position]
      properties:
        url: { type: string, format: uri }
        alt: { type: string }
        position: { type: integer }
    CategoryRef:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
    Variant:
      type: object
      required: [id, sku, priceCents, stock, attributes]
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        priceCents: { type: integer, description: 'Cents. 1900 = $19.00' }
        compareAtPriceCents:
          type: integer
          nullable: true
        stock: { type: integer }
        attributes:
          type: object
          additionalProperties: true
          example: { size: M }
    ProductSummary:
      type: object
      required: [id, slug, name, category]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        shortDescription:
          type: string
          nullable: true
        priceFromCents:
          type: integer
          nullable: true
          description: Lowest variant price; null when the product has no variants
        image:
          allOf: [{ $ref: '#/components/schemas/Image' }]
          nullable: true
        category: { $ref: '#/components/schemas/CategoryRef' }
    ProductDetail:
      type: object
      required: [id, slug, name, description, category, images, variants, inStock]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription:
          type: string
          nullable: true
        category: { $ref: '#/components/schemas/CategoryRef' }
        images:
          type: array
          items: { $ref: '#/components/schemas/Image' }
        variants:
          type: array
          items: { $ref: '#/components/schemas/Variant' }
        inStock: { type: boolean }
    PagedProducts:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ProductSummary' }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 12 }
        totalElements: { type: integer, example: 342 }
        totalPages: { type: integer, example: 29 }
...